pyproject.toml
[project]
name = "inspect-kathara"
version = "0.1.0"
description = "Kathara network lab integration for Inspect AI evaluations"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
authors = [{ name = "Open Telco Contributors" }]
keywords = ["inspect-ai", "kathara", "network", "evaluation", "ai", "benchmark"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking",
]

dependencies = [
    "inspect-ai>=0.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/otelcos/inspect-kathara"
Repository = "https://github.com/otelcos/inspect-kathara"
Documentation = "https://github.com/otelcos/inspect-kathara#readme"

[project.entry-points.inspect_ai]
kathara = "inspect_kathara._registry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inspect_kathara"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
